requirements/base.txt
django==1.6.1

# Configuration
django-configurations==0.7
django-cache-url==0.6.0

# Database
dj-database-url==0.2.2
psycopg2==2.5.1

# Models
South==0.8.4
django-model-utils==1.5.0

# Utils
django-reversion==1.8.0
django-extensions==1.2.5
git+https://github.com/django-mptt/django-mptt.git#egg=mptt

# Pydot Py3 (requires mercurial)
hg+https://bitbucket.org/prologic/pydot#egg=pydot

# API
djangorestframework==2.3.10
django-filter==0.7
markdown==2.3.1
